requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
dnspython>=2.6.0
pandas>=2.0.0
anthropic>=0.39.0
